split to F2Polynomial.cs
